backend/foodgram/requirements.txt
Django==2.2.19
djangorestframework==3.13.1
djoser==2.1.0
python-dotenv==0.20.0
webcolors==1.11.1
psycopg2-binary==2.8.6
pillow==9.1.0
django-filter==21.1
gunicorn==20.1.0
pytz==2022.1
sqlparse==0.4.2
asgiref==3.5.0